+++ Prg_UI/Wins/WinMenus/Taarif/TDETA_HES_SHEET4.xaml.cs
@@ -472,40 +472,38 @@ namespace Wins.WinMenus.Taarif
                                             var _hes = item.GetType().GetProperty("HESAB").GetValue(item);
 
 
-                                            //var RST = dbms.DoGetDataSQL<string?>("SELECT HES FROM DEED_DTL WHERE HES LIKE '" + N_KOL + "-" + NUMBER + "-" + TNUMBER + "-" + _tnumber2 + "-" + _tnumber3 + "-" + _tnumber4 + "%'").ToList();
-                                            //if (RST.Count > 0)
-                                            //{
-                                            //    e.Handled = true;
-
-                                            //    ErrosMessages.Add(new MsgModel { MessageText_U = $"حساب داراي گردش مي باشد قابل حذف نيست {_hes}" });
-                                            //}
-                                            //else
+                                            try
                                             {
-                                                try
+                                                var gerdeshCount = dbms.DoGetDataSQL<int>($@"SELECT COUNT(*) FROM dbo.DEED_DTL WHERE HES_K = {N_KOL} AND HES_M = {NUMBER} AND HES_T = {TNUMBER} AND HES_T2 = {_tnumber2} AND HES_T3 = {_tnumber3} AND HES_T4 = {_tnumber4}").FirstOrDefault();
+                                                if (gerdeshCount > 0)
                                                 {
-                                                    IsDeletedSomething = true;
-
+                                                    e.Handled = true;
+                                                    ErrosMessages.Add(new MsgModel { MessageText_U = $"این حساب دارای گردش است و نمیتوان آنرا حذف کرد!" });
+                                                }
+                                                else
+                                                {
                                                     ESLAH_ROW((int?)_tnumber4);
 
                                                     dbms.DoExecuteSQL($@" DELETE FROM dbo.TDETA_HES4 WHERE IDD = {_idd} ");
+                                                    IsDeletedSomething = true;
                                                 }
-                                                catch (SqlException ex)
+                                            }
+                                            catch (SqlException ex)
+                                            {
+                                                if (ex.Number == 547)
                                                 {
-                                                    if (ex.Number == 547)
-                                                    {
-                                                        e.Handled = true;
+                                                    e.Handled = true;
 
-                                                        ErrosMessages.Add(new MsgModel { MessageText_U = $"این حساب دارای گردش است و نمیتوان آنرا حذف کرد!" });
-                                                    }
-                                                    else
-                                                    {
-                                                        ErrosMessages.Add(new MsgModel { MessageText_U = "حذف به دلیل خطا در بروز پایگاه داده انجام نشد!" });
-                                                    }
+                                                    ErrosMessages.Add(new MsgModel { MessageText_U = $"این حساب دارای گردش است و نمیتوان آنرا حذف کرد!" });
                                                 }
-                                                catch (Exception)
+                                                else
                                                 {
-                                                    ErrosMessages.Add(new MsgModel { MessageText_U = "خطا در انجام عملیات حذف!" });
+                                                    ErrosMessages.Add(new MsgModel { MessageText_U = "حذف به دلیل خطا در بروز پایگاه داده انجام نشد!" });
                                                 }
+                                            }
+                                            catch (Exception)
+                                            {
+                                                ErrosMessages.Add(new MsgModel { MessageText_U = "خطا در انجام عملیات حذف!" });
                                             }
                                         }
                                     }
